moved shoot search and sort logic from store to list component
